requirements/dev.txt
-r production.txt

backcall==0.1.0
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.2.0
django-extensions==2.0.7
django-modeladmin-reorder==0.3.1
django-nested-admin==3.0.21
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
kombu==4.2.1
MarkupSafe==1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.6.4
python-monkey-business==1.0.0
pytz==2018.4
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
